New value-object Client was added.  Corresponding changes in dependent classes.
